Add Backbone collection and model!
Yay!
The background page creates a collection and fetches its content from
localstorage. Then scrapes the queue page and adds or updates models,
which are automatically saved to localstorage.
The popup page creates a Queue collection and fetches from localstorage
as well. This collection is used to create the elements.